Moved service out of hybrid cloud
